requirements.txt
streamlit
google-generativeai
sqlite3-0611
python-dotenv
pandas